Nuisance parameter estimation via honest random forest

This function estimates different nuisance parameters using the honest random forest implementation of the 'grf' package

Usage

NuPa_honest_forest(
  NuPa = c("Y.hat", "Y.hat.d", "Y.hat.z", "D.hat", "D.hat.z", "Z.hat"),
  X,
  Y = NULL,
  D = NULL,
  Z = NULL,
  n_cf_folds = 5,
  n_reps = 1,
  cluster = NULL,
  progress = FALSE,
  ...
)

Arguments

NuPa

String vector specifying the nuisance parameters to be estimated. Currently supported: c("Y.hat","Y.hat.d","Y.hat.z","D.hat","D.hat.z","Z.hat")

X

Covariate matrix with N rows and p columns.

Y

Optional numeric vector containing the outcome variable.

D

Optional binary treatment variable.

Z

Optional binary instrumental variable.

n_cf_folds

Number of cross-fitting folds. Default is 5.

n_reps

Number of repetitions of cross-fitting. Default is 1.

cluster

Optional vector of cluster variable if cross-fitting should account for clusters.

progress

If TRUE, progress of nuisance parameter estimation reported.

...

Options passed to the regression_forest.

Value

List of two lists.

predictions contains the requested nuisance parameters

smoothers contains the smoother matrices of requested outcome nuisance parameters

cf_mat Array of dimension n_reps x N x n_cf_folds storing indicators representing the folds used in estimation.

References

Wager, S., & Athey, S. (2018). Estimation and inference of heterogeneous treatment effects using random forests. Journal of the American Statistical Association, 113(523), 1228-1242.