summary method for class outcome_weights

Calculates several summary measures of potentially many outcome weights.

Usage

# S3 method for class 'get_outcome_weights'
summary(object, quiet = FALSE, digits = 4, epsilon = 1e-04, ...)

Arguments

object

get_outcome_weights object.

quiet

If TRUE, results are passed but not printed.

digits

Number of digits to be displayed. Default 4.

epsilon

Threshold below which in absolute values non-zero but small values should be displayed as < ...

...

further arguments passed to printCoefmat

Value

3D-array of dimension

c("Control","Treated") x

number of point estimates x

c("Minimum weight","Maximum weight","% Negative","Sum largest 10%","Sum of weights","Sum of absolute weights")